+++ src/vector/blas/axpy.cpp
@@ -16,7 +16,7 @@ namespace dd_avx{
 			get_isie(y.size(), is, ie);
 			reg alpha_hi = broadcast(alpha.x[0]);
 			reg alpha_lo = broadcast(alpha.x[1]);
-			for(i = is; (int)i < (int)(ie-SIMD_Length+1); i += SIMD_Length){
+			for(i = is; i < (ie-SIMD_Length+1); i += SIMD_Length){
 
 				reg x_hi = load(x.hi[i]);
 				reg x_lo = load(x.lo[i]);
@@ -48,7 +48,7 @@ namespace dd_avx{
 			get_isie(y.size(), is, ie);
 			reg alpha_hi = broadcast(alpha.x[0]);
 			reg alpha_lo = broadcast(alpha.x[1]);
-			for(i = is; (int)i < (int)(ie-SIMD_Length+1); i += SIMD_Length){
+			for(i = is; i < (ie-SIMD_Length+1); i += SIMD_Length){
 
 				reg x_hi = load(x.data()[i]);
 
@@ -79,7 +79,7 @@ namespace dd_avx{
 			get_isie(y.size(), is, ie);
 			reg alpha_hi = broadcast(alpha.x[0]);
 			reg alpha_lo = broadcast(alpha.x[1]);
-			for(i = is; (int)i < (int)(ie-SIMD_Length+1); i += SIMD_Length){
+			for(i = is; i < (ie-SIMD_Length+1); i += SIMD_Length){
 
 				reg x_hi = load(x.hi[i]);
 				reg x_lo = load(x.lo[i]);
@@ -110,7 +110,7 @@ namespace dd_avx{
 			get_isie(y.size(), is, ie);
 			reg alpha_hi = broadcast(alpha.x[0]);
 			reg alpha_lo = broadcast(0.0);
-			for(i = is; (int)i < (int)(ie-SIMD_Length+1); i += SIMD_Length){
+			for(i = is; i < (ie-SIMD_Length+1); i += SIMD_Length){
 
 				reg x_hi = load(x.data()[i]);
 
@@ -141,7 +141,7 @@ namespace dd_avx{
 			get_isie(y.size(), is, ie);
 			reg alpha_hi = broadcast(alpha);
 			reg alpha_lo = broadcast(0.0);
-			for(i = is; (int)i < (int)(ie-SIMD_Length+1); i += SIMD_Length){
+			for(i = is; i < (ie-SIMD_Length+1); i += SIMD_Length){
 
 				reg x_hi = load(x.hi[i]);
 				reg x_lo = load(x.lo[i]);
@@ -173,7 +173,7 @@ namespace dd_avx{
 			get_isie(y.size(), is, ie);
 			reg alpha_hi = broadcast(alpha);
 			reg alpha_lo = broadcast(0.0);
-			for(i = is; (int)i < (int)(ie-SIMD_Length+1); i += SIMD_Length){
+			for(i = is; i < (ie-SIMD_Length+1); i += SIMD_Length){
 
 				reg x_hi = load(x.data()[i]);
 				reg x_lo = regs.zeros;
@@ -205,7 +205,7 @@ namespace dd_avx{
 			get_isie(y.size(), is, ie);
 			reg alpha_hi = broadcast(alpha);
 			reg alpha_lo = broadcast(0.0);
-			for(i = is; (int)i < (int)(ie-SIMD_Length+1); i += SIMD_Length){
+			for(i = is; i < (ie-SIMD_Length+1); i += SIMD_Length){
 
 				reg x_hi = load(x.hi[i]);
 				reg x_lo = load(x.lo[i]);
